GenAi Projects/SpeakCoachAI/requirement.txt
openai
python-dotenv
SpeechRecognition
pyttsx3
pydub
gtts
fastapi
uvicorn
speechrecognition
pydantic
python-dotenv
google-cloud-texttospeech
google-cloud-speech
numpy
